requirements.txt
flask
google-cloud-ndb
appengine-python-standard
python-dotenv
